chore: organize redux files
