+++ src/components/Auth.tsx
@@ -1,6 +1,6 @@
 import { useDispatch } from 'react-redux'
 import { toast } from 'sonner'
-import { authorize } from '@/store/auth/authSlice'
+import { authorize } from '@/redux/slices/authSlice'
 import { Button } from '@/components/ui/button'
 
 export function Auth() {
