+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
 import App from './App.tsx'
-import { store } from '@/store/store'
+import { store } from '@/redux'
 
 import './index.css'
